add problems logging in links
